\name{portfolioGui}


\alias{guiPortfolio}  
\alias{portfolioGui} 
\alias{gui}


\title{Portfolio GUI}


\description{
  
    A portfolio Graphical user interface.
    
}
    

\usage{
showMainWindow()
getSelectedTreeViewEntry(treeView) 
listPortfolios()
showNewWindow(input)
showEditWindow0(button)
showEditWindow1(button)
showEditWindow(portfolioName0)
createPortfolio(input)
showPrintWindow(input2)
destroyWindow(input)
closeApp(input)
plotPortfolio(input)
getGladePath()
gui()
}


\arguments{

    \item{treeView}{
        ...
        }
    \item{input, input2}{
        ...
        }
    \item{button}{
        ...
        }
    \item{portfolioName0}{
        ...
        }

}


%\references{}


\examples{
## gui -  
   
}


\keyword{models}

